Use all diagnostic analyzers when running analyzer's tests (#2327)
* Use all analyzers for each linker's test run

* Remove SimpleAnalyzerOptions from CSharpAnalyzerVerifier
